Rename Sidecar and PSC to rev-a
