+++ task/power/src/main.rs
@@ -261,7 +261,7 @@ fn get_state() -> PowerState {
     }
 }
 
-#[cfg(target_board = "sidecar-1")]
+#[cfg(target_board = "sidecar-a")]
 fn controllers() -> [PowerController; 15] {
     let task = I2C.get_task_id();
 
@@ -284,7 +284,7 @@ fn controllers() -> [PowerController; 15] {
     ]
 }
 
-#[cfg(target_board = "sidecar-1")]
+#[cfg(target_board = "sidecar-a")]
 fn get_state() -> PowerState {
     task_slot!(SEQUENCER, sequencer);
