+++ task/thermal/src/bsp/sidecar_a.rs
