+++ test/tests-psc/app.toml
@@ -1,7 +1,7 @@
 name = "tests-psc"
 target = "thumbv7em-none-eabihf"
 chip = "../../chips/stm32h7"
-board = "psc-1"
+board = "psc-a"
 stacksize = 896
 
 [kernel]
